PDFBOX-1663: avoid ArrayOutOfBounds-Exception when dealing with monospaced fonts, changed width values from float to int
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1512661 13f79535-47bb-0310-9956-ffa450edef68

+++ preflight/src/main/java/org/apache/pdfbox/preflight/font/Type3FontValidator.java
@@ -276,7 +276,7 @@ public class Type3FontValidator extends FontValidator<Type3Container>
      */
     private void checkCharProcsAndMetrics() throws ValidationException
     {
-        List<Float> widths = font.getWidths();
+        List<Integer> widths = font.getWidths();
         if (widths == null || widths.isEmpty())
         {
             this.fontContainer.push(new ValidationError(ERROR_FONTS_DICTIONARY_INVALID,
@@ -460,7 +460,8 @@ public class Type3FontValidator extends FontValidator<Type3Container>
                     }
                     catch (IOException e)
                     {
-                        context.addValidationError(new ValidationError(PreflightConstants.ERROR_FONTS_DAMAGED, "Unable to valid the Type3 : " + e.getMessage()));
+                        context.addValidationError(new ValidationError(PreflightConstants.ERROR_FONTS_DAMAGED,
+                                "Unable to valid the Type3 : " + e.getMessage()));
                     }
                 }
             }
